将建筑系统改成建筑连续, 增加网格可视化测试工具

+++ Assets/Codes/Systems/BuildingSystem/BuildingData.cs
@@ -9,8 +9,8 @@ namespace Assets.Codes.Systems.BuildingSystem
     public class BuildingData
     {
         public string Name;
-        public int Width;
-        public int Height;
+        public float Width;
+        public float Height;
         public int Cost;
 
     }
